貯金目標の追加・編集・削除のテスト

+++ MoneyLogAPI/src/main/java/com/example/form/DeleteSavingTargetForm.java
@@ -1,20 +1,19 @@
 package com.example.form;
 
-import javax.validation.constraints.NotNull;
-
 import com.example.common.message.ValidatingMessage;
+
+import javax.validation.constraints.NotNull;
 
 public class DeleteSavingTargetForm extends form {
 
-	@NotNull(message = ValidatingMessage.ID_EMPTY_ERROR)
-	private Integer savingTargetId;
+    @NotNull(message = ValidatingMessage.ID_EMPTY_ERROR)
+    private Long savingTargetId;
 
-	public Integer getSavingTargetId() {
-		return savingTargetId;
-	}
+    public Long getSavingTargetId() {
+        return savingTargetId;
+    }
 
-	public void setSavingTargetId(Integer savingTargetId) {
-		this.savingTargetId = savingTargetId;
-	}
-
+    public void setSavingTargetId(Long savingTargetId) {
+        this.savingTargetId = savingTargetId;
+    }
 }
